UML scheme first revision
A UML scheme has been added.
There's alot of dependencies and interfaces.
Dia makes it sort of hard to see implementations of interfaces because
of the graphical style it used but I guess that's something you have to
live with.
Modifiyability and updatablility is done through the UpdatableComponent
interface and used the same name for RO and RW even though the API uses
different names. This is to have to rewrite as little code as possible.
Some methods are still missing such as class conversion from
ProcessedOutGoingSMS to LoggedOutgoingSMS