Add contact class and tests for it
The contact class the class to use when a contact hasn't been added to
the API yet. It has support to change all the attributes that the API
supports.

Next work will be done on the processed contact which will be named
ProcessedContact and implements the UpdatabaleComponent interface.
Comments hasn't been added to the class but will be in the near future
together with PHPDoc support.