Recipients use Factory class for construction
The static create* functions that were in the Contact class has now been
moved to a RecipientFactory this is so that the Recipeint classes can be
overwritten/extended in the future either by IP1 or by third party.
Unit Tests have also been altered to support this.
JSON input output tests (such as ContactTest::testCreateFromStdClass)
now compare against stdClass objects instead of JSON strings in so that
the order of the attributes in the JSON doesn't matter.

Signed-off-by: Hannes Kindströmmer <hannes@kindstrommer.se>