Add ProcessedContact and PHPDoc
Added a class called ProcessedContact which is a child of the Contact
class. The ProcessedContact class is what you will get in return when
adding a Contact to the API or when getting previously added contacts.
Tests for this class has also been added together with PHPDoc comments
for both the Contact and ProcessedContact class.

Signed-off-by: Hannes Kindströmmer <hannes@kindstrommer.se>