Implement MembershipRelation to ProcessedContact
Other than implementing MembershipRelation to ProcessedContact this also
contains functions for getting ProcessedContact(s) and ProcessedGroup(s)
that a Group has memberships for and is a member of respectivly.
The quality of life function has made it's way to RecipientFactory that
takes an array of stdClasses (containing group data) and turns it into
an array of ProcessedGroup. It's not much in terms of code but it will
lessen the lines of code required when dealing with arrays of
stdClasses.

Signed-off-by: Hannes Kindströmmer <hannes@kindstrommer.se>