Remove return type from Communicator::add method
The return type of the Communicator::add method has been removed as the return type can be as mentioned in the PHPDocs 'ProcessedComponentInterface ProcessedContact, ProcessedGroup, PrcessedMembership or a ClassValidatinArray' all the return tyes except ClassValidationArray implements ProcessedComponentInterface and ClassValidationArray will not implement that interface either because it doesn't work that way.
As said what is returned by the method in question is still mentioned in the PHPdocs.

Signed-off-by: Hannes Kindströmmer <hannes@kindstrommer.se>